added testing
