+++ tests/test_grad_accum.py
@@ -1,69 +1,12 @@
-import torch
-import os
-from torch import optim
-import torch.multiprocessing as mp
 import copy
 
-from open_lm.distributed import init_distributed_device
-from open_lm.train import train_one_epoch
-from open_lm.main import random_seed
-from open_lm.model import create_model
-from open_lm.data import get_data
-from open_lm.scheduler import cosine_lr
-from tests.shared import MockTrainArgs
-
+import torch
+import torch.multiprocessing as mp
 from torch.distributed.fsdp import FullyShardedDataParallel as FSDP
 
-
-def create_grad_acc_fixtures():
-    # Setup data, optimizer, and other basic settings
-    args = MockTrainArgs("open_lm_11m")
-
-    # only want to look at one batch
-    args.train_num_samples = args.batch_size
-
-    # increase learning rate and remove warmup for maximize change to model weights
-    args.lr = 2
-    args.warmup = 0
-
-    # create base models
-    random_seed()
-    model = create_model(args).to(args.device)
-
-    # create dataloader
-    data = get_data(
-        args,
-        epoch=0,
-        tokenizer=None,
-        skip_train=False,
-    )
-
-    # create optimizer
-    named_parameters = list(model.named_parameters())
-    params = [p for _, p in named_parameters if p.requires_grad]
-    optimizer = optim.AdamW(
-        [
-            {"params": params, "weight_decay": args.wd},
-        ],
-        lr=args.lr,
-        betas=(args.beta1, args.beta2),
-        eps=args.eps,
-    )
-
-    # create scheduler
-    scheduler = cosine_lr(
-        optimizer,
-        args.lr,
-        args.warmup,
-        10,
-        args.lr_cooldown_end,
-        args.force_min_lr,
-    )
-
-    # create loss
-    loss = torch.nn.CrossEntropyLoss()
-
-    return args, model, data, optimizer, scheduler, loss
+from open_lm.train import train_one_epoch
+from open_lm.main import random_seed
+from tests.shared import create_train_fixtures
 
 
 def _grad_acc_helper(test_fsdp, accs=[2, 1], threshold=1e-7):
@@ -92,7 +35,7 @@ def _grad_acc_helper_fsdp(rank, world_size, accs, threshold):
 
 
 def _grad_acc_helper_single(test_fsdp, accs=[2, 1], threshold=1e-7):
-    args, model, data, optimizer, scheduler, loss = create_grad_acc_fixtures()
+    args, model, data, optimizer, scheduler, loss = create_train_fixtures()
 
     if test_fsdp:
         args.fsdp = True
